Move the local images hack into it's own app so it only runs on dev
